Update custom-theme.md
I tried creating my own theme and followed the documentation, but let my IDE auto-import the variables because the page didn't include them. 

However, I was stuck trying to implement a different interface than the one in the page. That's when I realised that I accidentally imported either an old version of the Theme or a different version, either way I think adding the import would be useful.